@@ -6,6 +6,7 @@ import traceback
 import base64
 from uuid import UUID
 
+from django.core.exceptions import RequestDataTooBig
 from django.urls import resolve, Resolver404
 
 from silk import models
@@ -24,6 +25,16 @@ content_type_form = ['multipart/form-data',
 content_type_html = ['text/html']
 content_type_css = ['text/css']
 
+
+def _get_response_headers(response):
+    """
+    Django 3.2 (more specifically, commit bcc2befd0e9c1885e45b46d0b0bcdc11def8b249) broke the usage of _headers, which
+    were turned into a public interface, so we need this compatibility wrapper.
+    """
+    try:
+        return response.headers
+    except AttributeError:
+        return response._headers
 
 class DefaultEncoder(json.JSONEncoder):
     def default(self, o):
@@ -92,7 +103,7 @@ class RequestModelFactory(object):
             except KeyError:
                 pass
 
-        return json.dumps(headers, cls=DefaultEncoder)
+        return json.dumps(headers, cls=DefaultEncoder, ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
 
     def _mask_credentials(self, body):
         """
@@ -126,7 +137,7 @@ class RequestModelFactory(object):
             except Exception:
                 Logger.debug('{}'.format(str(e)))
         else:
-            body = json.dumps(replace_pattern_values(json_body))
+            body = json.dumps(replace_pattern_values(json_body), ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
 
         return body
 
@@ -138,17 +149,26 @@ class RequestModelFactory(object):
         body = ''
         if content_type in content_type_form:
             body = self.request.POST
-            body = json.dumps(dict(body), sort_keys=True, indent=4)
+            body = json.dumps(dict(body), sort_keys=True, indent=4
+                              , ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
         elif content_type in content_types_json:
             try:
-                body = json.dumps(json.loads(raw_body), sort_keys=True, indent=4)
+                body = json.dumps(json.loads(raw_body), sort_keys=True, indent=4
+                                  , ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
             except:
                 body = raw_body
         return body
 
     def body(self):
         content_type, char_set = self.content_type()
-        raw_body = self.request.body
+        try:
+            raw_body = self.request.body
+        except RequestDataTooBig:
+            raw_body = b"Raw body exceeds DATA_UPLOAD_MAX_MEMORY_SIZE, Silk is not showing file uploads."
+            body = self.request.POST.copy()
+            for k, v in self.request.FILES.items():
+                body.appendlist(k, v)
+            return body, raw_body
         if char_set:
             try:
                 raw_body = raw_body.decode(char_set)
@@ -211,7 +231,7 @@ class RequestModelFactory(object):
         encoded_query_params = ''
         if query_params:
             query_params_dict = dict(zip(query_params.keys(), query_params.values()))
-            encoded_query_params = json.dumps(query_params_dict)
+            encoded_query_params = json.dumps(query_params_dict, ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
         return encoded_query_params
 
     def view_name(self):
@@ -284,7 +304,8 @@ class ResponseModelFactory(object):
                     # and json.dumps(...) in python3
                     content = content.decode()
                 try:
-                    body = json.dumps(json.loads(content), sort_keys=True, indent=4)
+                    body = json.dumps(json.loads(content), sort_keys=True, indent=4
+                                      , ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII)
                 except (TypeError, ValueError):
                     Logger.warn(
                         'Response to request with pk %s has content type %s but was unable to parse it'
@@ -299,7 +320,7 @@ class ResponseModelFactory(object):
             % self.request.pk
         )
         b, content = self.body()
-        raw_headers = self.response._headers
+        raw_headers = _get_response_headers(self.response)
         headers = {}
         for k, v in raw_headers.items():
             try:
@@ -311,7 +332,7 @@ class ResponseModelFactory(object):
         silky_response = models.Response.objects.create(
             request_id=self.request.id,
             status_code=self.response.status_code,
-            encoded_headers=json.dumps(headers),
+            encoded_headers=json.dumps(headers, ensure_ascii=SilkyConfig().SILKY_JSON_ENSURE_ASCII),
             body=b
         )
 
